+++ web/components/issues/issue-layouts/list/root.tsx
@@ -36,10 +36,7 @@ export const ListLayout: FC = observer(() => {
         issueStore.updateIssueStructure(group_by, null, issue);
         issueDetailStore.updateIssue(workspaceSlug.toString(), issue.project, issue.id, issue);
       }
-      if (action === "delete") {
-        issueStore.deleteIssue(group_by, null, issue);
-        issueDetailStore.deleteIssue(workspaceSlug.toString(), issue.project, issue.id);
-      }
+      if (action === "delete") issueStore.deleteIssue(group_by, null, issue);
     },
     [issueStore, issueDetailStore, workspaceSlug]
   );
